backend/pyproject.toml
[project]
name = "matchmaster-backend"
version = "0.1.0"
description = "AI Recruitment Screening Tool — Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "redis>=5.2,<6",
    "openai>=1.58,<2",
    "httpx>=0.28,<1",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0,<2",
    "pymupdf>=1.25,<2",
    "python-docx>=1.1,<2",
    "jinja2>=3.1,<4",
    "numpy>=2.1,<3",
    "tiktoken>=0.8,<1",
    "sse-starlette>=2.2,<3",
    "pgvector>=0.3,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6.0,<7",
    "httpx>=0.28,<1",
    "ruff>=0.8,<1",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
